fixing NPE

+++ src/main/java/org/tdar/core/bean/coverage/LatitudeLongitudeBox.java
@@ -244,7 +244,7 @@ public class LatitudeLongitudeBox extends Persistable.Base implements HasResourc
         if (isOkayToShowExactLocation) {
             result = actualValue;
         }
-        if (!Objects.equals(actualValue.doubleValue(), obfuscatedValue.doubleValue())) {
+        if (!Objects.equals(actualValue, obfuscatedValue)) {
             setObfuscatedObjectDifferent(true);
         }
         return result;
